# ui/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "CatchupPolicy": {
        "enum": [
          "skip",
          "latest",
          "all"
        ],
        "type": "string"
      },
...
      "Flow": {
        "description": "A registered flow. Identity is `(project, name)`.",
        "properties": {
          "created_at": {
            "$ref": "#/components/schemas/i64"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "error": {
            "type": [
              "string",
              "null"
            ]
          },
          "external_id": {
            "$ref": "#/components/schemas/Id"
          },
          "group": {
            "description": "The group the flow is listed under, declared in Python with `group=`.\nNull means none was declared and the flow is grouped under its project;\nAPI responses carry the resolved value, so clients never apply that\nfallback themselves.",
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "last_seen_at": {
            "$ref": "#/components/schemas/i64"
          },
          "live": {
            "description": "Whether the running server has this flow registered from code.",
            "type": "boolean"
          },
          "module": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "options": {
            "type": "object"
          },
          "parameter_schema": {
            "type": "object"
          },
          "project": {
            "type": "string"
          },
          "source_dir": {
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": [
          "id",
          "external_id",
          "project",
          "name",
          "module",
          "source_dir",
          "created_at",
          "last_seen_at"
        ],
        "type": "object"
      },
...
      "FlowSummary": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Flow"
          },
          {
            "properties": {
              "batch_key": {
                "description": "The batch key of a keyed fan-in.",
                "type": [
                  "string",
                  "null"
                ]
              },
              "recent_runs": {
                "description": "Newest first: (run id, state type, state name, duration in microseconds\nor null while the run has none) of the last ten runs.",
                "items": {
                  "items": false,
                  "prefixItems": [
                    {
                      "format": "int64",
                      "type": "integer"
                    },
                    {
                      "type": "string"
                    },
                    {
                      "type": "string"
                    },
                    {
                      "format": "int64",
                      "type": [
                        "integer",
                        "null"
                      ]
                    }
                  ],
                  "type": "array"
                },
                "type": "array"
              },
              "schedules": {
                "items": {
                  "$ref": "#/components/schemas/ScheduleRow"
                },
                "type": "array"
              },
              "triggered_by": {
                "description": "The flow this one runs after, if any (the first upstream).",
                "type": [
                  "string",
                  "null"
                ]
              },
              "triggers": {
                "description": "Flows that run after this one.",
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              "upstreams": {
                "description": "Every upstream of a fan-in flow (empty without `after=`).",
                "items": {
                  "type": "string"
                },
                "type": "array"
              }
            },
            "required": [
              "recent_runs",
              "triggers",
              "schedules"
            ],
            "type": "object"
          }
        ]
      },
...
      "Id": {
        "type": "string"
      },
...
      "RouteSpec": {
        "properties": {
          "id": {
            "minimum": 0,
            "type": "integer"
          },
          "method": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "source": {
            "description": "Where the handler is defined (module and function), for the settings page.",
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "id",
          "method",
          "path"
        ],
        "type": "object"
      },
...
      "Schedule": {
        "oneOf": [
          {
            "properties": {
              "cron": {
                "type": "string"
              },
              "day_or": {
                "type": "boolean"
              },
              "kind": {
                "enum": [
                  "cron"
                ],
                "type": "string"
              },
              "timezone": {
                "type": [
                  "string",
                  "null"
                ]
              }
            },
            "required": [
              "cron",
              "kind"
            ],
            "type": "object"
          },
          {
            "properties": {
              "anchor": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/i64",
                    "description": "Anchor time in microseconds UTC; defaults to the schedule's creation."
                  }
                ]
              },
              "interval": {
                "description": "Seconds between fires.",
                "format": "double",
                "type": "number"
              },
              "kind": {
                "enum": [
                  "interval"
                ],
                "type": "string"
              },
              "timezone": {
                "type": [
                  "string",
                  "null"
                ]
              }
            },
            "required": [
              "interval",
              "kind"
            ],
            "type": "object"
          },
          {
            "properties": {
              "kind": {
                "enum": [
                  "rrule"
                ],
                "type": "string"
              },
              "rrule": {
                "type": "string"
              },
              "timezone": {
                "type": [
                  "string",
                  "null"
                ]
              }
            },
            "required": [
              "rrule",
              "kind"
            ],
            "type": "object"
          }
        ]
      },
...
      "ScheduleRow": {
        "description": "Placeholders for later phases; defined now so the schema and API types\nare stable from the start.\nA stored schedule row: the schedule itself plus policy and bookkeeping.",
        "properties": {
          "active": {
            "type": "boolean"
          },
          "catchup": {
            "$ref": "#/components/schemas/CatchupPolicy"
          },
          "catchup_max": {
            "format": "int64",
            "type": "integer"
          },
          "code_key": {
            "type": [
              "string",
              "null"
            ]
          },
          "created_at": {
            "$ref": "#/components/schemas/i64"
          },
          "external_id": {
            "$ref": "#/components/schemas/Id"
          },
          "flow_id": {
            "format": "int64",
            "type": "integer"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "next_fire": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/i64",
                "description": "Next fire that will run, computed by the scheduler (not stored)."
              }
            ]
          },
          "paused_reason": {
            "type": [
              "string",
              "null"
            ]
          },
          "paused_until": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/i64"
              }
            ]
          },
          "persist": {
            "type": "boolean"
          },
          "schedule": {
            "$ref": "#/components/schemas/Schedule"
          },
          "skipped": {
            "description": "Future fires skipped by a person, counted by the scheduler (not stored).",
            "format": "int64",
            "type": "integer"
          },
          "source": {
            "description": "`code` for schedules declared on the flow, `ui` for ones created in the\ninterface, `mcp` for ones created by an agent. Startup reconciliation\nsingles out `code` alone; every other value is left as it is.",
            "type": "string"
          },
          "updated_at": {
            "$ref": "#/components/schemas/i64"
          }
        },
        "required": [
          "id",
          "external_id",
          "flow_id",
          "schedule",
          "catchup",
          "catchup_max",
          "active",
          "source",
          "persist",
          "created_at",
          "updated_at"
        ],
        "type": "object"
      },
...
      "Settings": {
        "properties": {
          "catchup_default": {
            "type": "string"
          },
          "crash_retries_default": {
            "format": "int64",
            "type": "integer"
          },
          "custom_routes": {
            "items": {
              "$ref": "#/components/schemas/RouteSpec"
            },
            "type": "array"
          },
          "database_bytes": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          },
          "database_path": {
            "type": "string"
          },
          "email_configured": {
            "type": "boolean"
          },
          "engine_max_runs": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "engine_saturation_risk": {
            "type": "boolean"
          },
          "home": {
            "type": "string"
          },
          "host": {
            "type": "string"
          },
          "max_engines": {
            "minimum": 0,
            "type": "integer"
          },
          "pid": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "port": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "resources": {
            "type": "object"
          },
          "retain_days": {
            "format": "int64",
            "type": "integer"
          },
          "saturation_flows": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "saturation_reason": {
            "type": [
              "string",
              "null"
            ]
          },
          "secret_key_missing": {
            "type": "boolean"
          },
          "secret_key_present": {
            "type": "boolean"
          },
          "served_dir": {
            "type": [
              "string",
              "null"
            ]
          },
          "title": {
            "description": "UI title in effect: `[ui] title` from cereyan.toml, or `cereyan`.",
            "type": "string"
          },
          "version": {
            "type": "string"
          },
          "wal_bytes": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "host",
          "port",
          "pid",
          "version",
          "title",
          "home",
          "database_path",
          "database_bytes",
          "wal_bytes",
          "resources",
          "max_engines",
          "engine_max_runs",
          "crash_retries_default",
          "catchup_default",
          "retain_days",
          "email_configured",
          "secret_key_present",
          "secret_key_missing",
          "custom_routes",
          "engine_saturation_risk",
          "saturation_flows"
        ],
        "type": "object"
      },
      "SettingsPatch": {
        "properties": {
          "crash_retries": {
            "format": "int64",
            "type": [
              "integer",
              "null"
            ]
          },
          "resources": {
            "additionalProperties": {
              "format": "double",
              "type": "number"
            },
            "propertyNames": {
              "type": "string"
            },
            "type": [
              "object",
              "null"
            ]
          },
          "retain_days": {
            "format": "int64",
            "type": [
              "integer",
              "null"
            ]
          },
          "title": {
            "description": "UI title; an empty string removes `[ui] title` and restores `cereyan`.",
            "type": [
              "string",
              "null"
            ]
          }
        },
        "type": "object"
      },
...
      "i64": {
        "format": "int64",
        "type": "integer"
      }
    }
  },
  "info": {
    "description": "Local-first pipeline orchestrator API",
    "license": {
      "identifier": "MIT",
      "name": "MIT"
    },
    "title": "cereyan",
    "version": "1.14.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/flows": {
      "get": {
        "operationId": "list_flows",
        "parameters": [
          {
            "in": "query",
            "name": "project",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FlowSummary"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "flows"
        ]
      }
    },
...
    "/api/settings": {
      "get": {
        "operationId": "get_settings",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Settings"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "settings"
        ]
      },
      "patch": {
        "operationId": "patch_settings",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Settings"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "settings"
        ]
      }
    },
...
  }
}